+++ src/updater.h
@@ -3,8 +3,8 @@
 
 #include <windows.h>
 
-#define PERSISTENT_CALCULATOR_VERSION_A "2.0.0"
-#define PERSISTENT_CALCULATOR_VERSION_W L"2.0.0"
+#define PERSISTENT_CALCULATOR_VERSION_A "2.0.1"
+#define PERSISTENT_CALCULATOR_VERSION_W L"2.0.1"
 
 int updater_apply_if_requested(void);
 void updater_initialize(HWND owner);